+++ src/Exceptions/StateNotFoundException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Caner\StateMachine\Exceptions;
+
+use RuntimeException;
+
+class StateNotFoundException extends RuntimeException
+{
+}