+++ src/Interfaces/BaseGuardInterface.php
@@ -2,20 +2,11 @@
 
 namespace Caner\StateMachine\Interfaces;
 
-use Caner\StateMachine\Concerns\BaseStateMachine;
-use Illuminate\Http\Request;
-
 interface BaseGuardInterface
 {
-    public function __construct(
-        BaseStateMachine $stateMachine,
-        ?Request $request = null,
-        array $data = []
-    );
-
     public function check(): self;
 
     public function getRequestData(): array;
 
-    public function completed();
+    public function completed(): void;
 }